+++ NEWS.md
@@ -10,6 +10,8 @@
 
 * New function `vetiver_deploy_rsconnect()` to deploy to RStudio Connect.
 
+* Added R Markdown template for [Model Card](https://doi.org/10.1145/3287560.3287596) for responsible, transparent model reporting (#62).
+
 # vetiver 0.1.0
 
 * Initial CRAN release of package.